Boost almost all code from DefaultViewController to its previous base class, ViewController, so that it is shared with macOS
